+++ lab_3/result.json
@@ -1,4 +1,4 @@
 {
-  "variant": "1000",
-  "checksum": "b3cef7aba7dc0f1680189b41b3f7c800"
+  "variant": "27",
+  "checksum": "f2b80fe52cd8d0648338f9d2ac7b9303"
 }
