--- result.json
@@ -1,4 +1,0 @@
-{
-  "variant": "27",
-  "checksum": "f2b80fe52cd8d0648338f9d2ac7b9303"
-}
